+++ tests/test_questions.py
@@ -9,89 +9,7 @@ from flask_testing import TestCase
 from api.database.connect import conn, cur
 import os
 from api.manage import tables
-
-
-
-
-
-class Base(TestCase):
-    """contains config for testing"""
-
-    def create_app(self):
-        """sets config to testing"""
-        self.app = create_app('testing')
-        return self.app
-
-    def setUp(self):
-        tables("create")
-        self.app_context = self.app.app_context()
-        self.app_context.push()
-        self.test_client = self.app.test_client()
-        self.sample_data1 = {
-            'description': 'A tuple is a python data structure'
-        }
-
-        self.sample_data2 = {
-            'title': 'What is a title?'
-        }
-
-        self.sample_data3 = {
-            'title': '',
-            'description': 'A tuple is a python data structure'
-
-        }
-
-        self.sample_data4 = {
-            'title': 'What is a title?',
-            'description': ''
-
-        }
-
-        self.sample_data5 = {
-            'title': '12345',
-            'description': 'A tuple is a python data structure'
-
-        }
-
-        self.sample_data6 = {
-            'title': 'What is a title?',
-            'description': '12345'
-
-        }
-
-        self.sample_data7 = {
-
-            'title': 'What is a tuple?',
-            'description': 'A tuple is a python data structure'
-        }
-
-        self.signup_details = {
-            "first_name" : "John",
-            "last_name" : "Doe",
-            "username" : "josdhndoe",
-            "email" : "johndsdoe@gmail.com",
-            "password" : "absdcd1234"
-            } 
-
-        self.login_details = {            
-            "username" : "josdhndoe",
-            "password" : "absdcd1234"           
-            } 
-
-        self.client.post(
-            '/api/v1/auth/signup',
-            data=json.dumps(self.signup_details),
-            content_type='application/json')
-
-        self.que = self.client.post(
-            '/api/v1/auth/login',
-            data=json.dumps(self.login_details),
-            content_type='application/json')
-
-
-    def tearDown(self):
-        self.app_context.pop()    
-        tables("drop")
+from tests.base import Base
 
 class TestApp(Base):
